// frontend/src/services/household_debt_api/api.ts
export interface CountyDebt {
	/**
	 * County
	 */
	county: string
	/**
	 * Average debt ratio
	 */
	average_debt: number
	/**
	 * FIPS code for the County
	 */
	fips_code: string
}
export interface CountyExtended {
	/**
	 * County Name
	 */
	name: string
	/**
	 * County Fips Code
	 */
	fips_code: string
	/**
	 * State Code for County
	 */
	state_code: string
}
export interface Debt {
	/**
	 * Date and time of debt
	 */
	date: string
	/**
	 * Lower bound of debt
	 */
	low: number
	high: number | null
}
export interface DebtExtended {
	/**
	 * Date and time of debt
	 */
	date: string
	/**
	 * Lower bound of debt
	 */
	low: number
	high: number | null
	/**
	 * County of Debt
	 */
	county: string
	/**
	 * State of County with Debt
	 */
	state: string
	/**
	 * FIPS code for the County
	 */
	fips_code: string
}
export interface HTTPValidationError {
	detail?: Array<ValidationError>
}
export interface State {
	/**
	 * State Name
	 */
	name: string
	/**
	 * State Code
	 */
	code: string
	/**
	 * State Fips Code
	 */
	fips_code: string
}
export interface StateDebt {
	/**
	 * State
	 */
	state: string
	/**
	 * Average debt ratio
	 */
	average_debt: number
	/**
	 * Number of Counties
	 */
	number_of_counties: number
	/**
	 * FIPS code for the State
	 */
	fips_code: string
}